Add ZooKeeper 3.5.2-alpha bundle
This bundle of ZooKeeper is needed for Windows, as 3.4.x doesn't compile
on Windows. Additionally, we have a patch which can applied to this
version that adds a CMake build system to ZooKeeper, unblocking our use
of Visual Studio 2017.

The 3.5.x series of ZooKeeper is "alpha" while 3.4.x is "stable." That
said, there is a 3.5.3 version still in RC, while 3.5.2 has been
released to RC and has a distribution available.

This tarball comes from the ZooKeeper official distributions, and as
such, includes some generated files necessary for building the C
bindings. A source-only tarball (like that which GitHub provides) is
insufficient, and would require additional build tools (ant and Java) as
well as additional steps to build.